api_service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
psycopg2-binary
pika
email-validator
